Struct librsync::Signature

pub struct Signature<R> {
    // some fields omitted
}

A struct to generate a signature.

This type takes a Read stream for the input from which compute the signatures, and implements another Read stream from which get the result.

Methods

impl<R: Read> Signature<BufReader<R>>

fn new(input: R) -> Result<Self>

Creates a new signature stream with default parameters.

This constructor takes an input stream for the file from which compute the signatures. Default options are used for the signature format: BLAKE2 for the hashing, 2048 bytes for the block length and full length for the strong signature size.

fn with_options(input: R, block_len: usize, strong_len: usize, sig_magic: SignatureType) -> Result<Self>

Creates a new signature stream by specifying custom parameters.

This constructor takes the input stream for the file from which compute the signatures, the size of checksum blocks as block_len parameter (larger values make the signature shorter and the delta longer), and the size of strong signatures in bytes as strong_len parameter. If it is non-zero the signature will be truncated to that amount of bytes. The last parameter specifies which version of the signature format to be used.

impl<R: BufRead> Signature<R>

fn with_buf_read(input: R, block_len: usize, strong_len: usize, sig_magic: SignatureType) -> Result<Self>

Creates a new signature stream by using a BufRead.

This constructor takes an already built BufRead instance. Prefer this constructor if you already have a BufRead as input stream, since it avoids wrapping the input stream into another BufRead instance. See with_options constructor for details on the other parameters.

fn into_inner(self) -> R

Unwraps this stream, returning the underlying input stream.